fix(tokenizer): TokenizersTokenizer -> HFTokenizer
https://github.com/pytorch/torchchat/issues/1251
Branch: TokenizersCpp-1251
